Add reusable faker provider for NHS number.
