+++ tests/fixtures/builders/repos/person.py
@@ -1,11 +1,11 @@
-import string
 from collections.abc import Sequence
 from datetime import date
-from random import choice, randint, shuffle
+from random import choice, shuffle
 from typing import Any, Literal, get_args
 
 from faker import Faker
-from faker.providers import BaseProvider
+
+from tests.fixtures.builders.providers import PersonDetailProvider
 
 Gender = Literal["0", "1"]
 
@@ -88,35 +88,3 @@ def person_rows_builder(  # noqa:PLR0913
 
     shuffle(rows)
     return rows
-
-
-class PersonDetailProvider(BaseProvider):
-    def icb(self) -> str | None:
-        if randint(0, 3):
-            return f"{choice(string.ascii_uppercase)}{choice(string.ascii_uppercase)}{choice(string.digits)}"
-        return None
-
-    def gp_practice(self) -> str | None:
-        if randint(0, 3):
-            return f"{choice(string.ascii_uppercase)}{randint(1, 99999):05}"
-        return None
-
-    def pcn(self) -> str | None:
-        if randint(0, 3):
-            return f"{choice(string.ascii_uppercase)}{randint(1, 99999):05}"
-        return None
-
-    def comissioning_region(self) -> str | None:
-        if randint(0, 3):
-            return f"{choice(string.ascii_uppercase)}{randint(1, 99):02}"
-        return None
-
-    def msoa(self) -> str | None:
-        if randint(0, 3):
-            return f"{choice(string.ascii_uppercase)}{randint(1, 99999999):08}"
-        return None
-
-    def lsoa(self) -> str | None:
-        if randint(0, 3):
-            return f"{choice(string.ascii_uppercase)}{randint(1, 99999999):08}"
-        return None
